radmod_namelist change
Added "NO3=1" as default.  Previously NO3 option was not in the namelist by default. If one set the NO3 option in the most_plasim_run_mod script but did not add NO3 manually to the radmod_namelist then the script could not change the namelist to include the NO3 value desired.  Adding the NO3=1 default fixes this, so that changes to NO3 value in run_mod script will actually take effect in the radmod_namelist without having to make sure NO3 is in that namelist manually.